Moved user-related types to core.identity
This commit moves several user-related types to core.identity, to
resolve a cyclic dependency.

See: gh-1024
