+++ embabel-agent-api/src/main/kotlin/com/embabel/chat/Chatbot.kt
@@ -16,7 +16,7 @@
 package com.embabel.chat
 
 import com.embabel.agent.api.channel.OutputChannel
-import com.embabel.agent.api.identity.User
+import com.embabel.agent.core.identity.User
 import com.embabel.agent.core.Budget
 
 /**
