+++ embabel-agent-api/src/test/kotlin/com/embabel/chat/ChatSessionTest.kt
@@ -19,7 +19,7 @@ import com.embabel.agent.api.channel.MessageOutputChannelEvent
 import com.embabel.agent.api.channel.OutputChannel
 import com.embabel.agent.api.channel.OutputChannelEvent
 import com.embabel.agent.api.reference.LlmReference
-import com.embabel.agent.api.identity.User
+import com.embabel.agent.core.identity.User
 import com.embabel.agent.core.hitl.ConfirmationRequest
 import com.embabel.chat.support.InMemoryConversation
 import org.assertj.core.api.Assertions.assertThat
