+++ embabel-agent-api/src/test/kotlin/com/embabel/chat/ConversationFactoryProviderTest.kt
@@ -15,7 +15,7 @@
  */
 package com.embabel.chat
 
-import com.embabel.agent.api.identity.User
+import com.embabel.agent.core.identity.User
 import com.embabel.chat.support.InMemoryConversation
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
